update beef API 18-11-2022

+++ tbeef/createToDb.php
@@ -10,16 +10,26 @@ $database = new Database();
 $db = $database->getConnection();
 $obj = new tbeef($db);
 $data = json_decode(file_get_contents("php://input"));
-$obj->beefFolder = $data->beefFolder;
-$obj->beefCode = $data->beefCode;
-$obj->fraction = $data->fraction;
-$obj->status = 0;
-$obj->classifyJudge = 0;
-$obj->createDate = Date("Y-m-d");
-if($obj->create()){
+$beefFolder = $data->beefFolder;
+if($obj->getFolderExist($beefFolder)==false){
+	$obj->beefFolder = $data->beefFolder;
+	$obj->beefCode = $data->beefCode;
+	$obj->fraction = $data->fraction;
+	$obj->status = 0;
+	$obj->classifyJudge = 0;
+	$obj->createDate = Date("Y-m-d");
+
+	if($obj->create()){
 		echo json_encode(array('message'=>true));
+	}
+	else{
+			echo json_encode(array('message'=>false));
+	}
 }
-else{
-		echo json_encode(array('message'=>false));
-}
+
+
+
+
+
+
 ?>
